requirements.txt
fastapi
uvicorn
python-telegram-bot
python-dotenv
motor
httpx